Update bgd-app.yaml
Changed the namespace to work with openshift gitops out of the box.
Install the operator - the operator automatically creates openshift-gitops namespace and installs argocd.  You have to put applications in that namespace in order for argocd to pick it up.  The docs show how to work around this by creating the spaces, but for demo purposes this is quicker than having to create new projects.